+++ src/main/java/com/gylang/gylangauthshirojpa/controller/SysDeptController.java
@@ -7,9 +7,12 @@ import com.gylang.gylangauthshirojpa.enums.ResultEnum;
 import com.gylang.gylangauthshirojpa.form.PageForm;
 import com.gylang.gylangauthshirojpa.service.SysDeptService;
 import com.gylang.gylangauthshirojpa.utils.CopyUtils;
+import com.gylang.gylangauthshirojpa.utils.JsonUtils;
 import com.gylang.gylangauthshirojpa.utils.LoginUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.util.CollectionUtils;
+import org.springframework.util.ObjectUtils;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -33,16 +36,21 @@ public class SysDeptController extends BaseController<SysDept> {
 
     @Override
     public Result save(@RequestBody SysDept sysDept) {
-
-
+        System.out.println(JsonUtils.obj2Json(sysDept));
         return Result.auto(null != sysDeptService.save(sysDept));
     }
 
     @Override
     public Result update(@RequestBody SysDept sysDept) {
 
-        SysDept dept = new SysDept();
-
+        if (null == sysDept.getId()) {
+            return Result.failure(ResultEnum.ID_NULL);
+        }
+        System.out.println(JsonUtils.obj2Json(sysDept));
+        SysDept dept = sysDeptService.findById(sysDept.getId());
+        if (ObjectUtils.isEmpty(dept)) {
+            return Result.failure(ResultEnum.RESULT_EMPTY);
+        }
         List<String> ignore = Arrays.asList("createTime", "createBy");
         CopyUtils.notNullAndOtherCopy(sysDept, dept, ignore);
 
@@ -51,7 +59,7 @@ public class SysDeptController extends BaseController<SysDept> {
 
     @Override
     public Result delete(@RequestBody List<SysDept> t) {
-
+        System.out.println(JsonUtils.obj2Json(t));
         List<Long> deptId = t.stream()
                 .filter(sysConfig -> null != sysConfig.getId())
                 .map(SysDept::getId).collect(Collectors.toList());
@@ -64,7 +72,9 @@ public class SysDeptController extends BaseController<SysDept> {
         List<SysDept> delete = sysDeptList.stream()
                 .filter(sysDept -> loginInfoDTO.getName().equals(sysDept.getCreateBy()))
                 .collect(Collectors.toList());
-
+        if (CollectionUtils.isEmpty(delete)) {
+            return Result.failure(ResultEnum.RESULT_EMPTY);
+        }
         return Result.auto(sysDeptService.delete(delete));
     }
 
@@ -72,5 +82,10 @@ public class SysDeptController extends BaseController<SysDept> {
     public Result findPage(@RequestBody @Valid PageForm pageForm) {
 
         return Result.success(sysDeptService.findPage(pageForm));
+    }
+
+    @GetMapping(value="/findTree")
+    public Result findTree() {
+        return Result.success(sysDeptService.findTree());
     }
 }
